requirements.txt
python = 3.10.6
Django = 4.1.7
djangorestframework = 3.14.0
django-cors-headers = 3.14.0
django-filter = 22.1
markdown = 3.4.1
psycopg2 = 2.9.5
